Implement FileHelper.kt to handle file operations
